Project source code
The project aims to develop a neural network model that generates descriptive captions for uploaded images. By training on a diverse dataset of image-caption pairs, the model will learn to recognize and describe various objects and scenes. The focus is on optimizing accuracy and efficiency, with the end goal of deploying the model in an application or API for real-world use